% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal.table.R
\name{bal.table}
\alias{bal.table}
\title{Compute the balance table for a ps.cont object.}
\usage{
bal.table(x, digits = 3, ...)
}
\arguments{
\item{x}{A `ps.cont` object}

\item{digits}{Number of digits to round to}

\item{...}{Additional arguments.}
}
\description{
Compute the balance table for a ps.cont object.
}
