% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ps.cont.R
\name{summary.ps.cont}
\alias{summary.ps.cont}
\title{Displays a useful description of a `ps.cont` object.}
\usage{
\method{summary}{ps.cont}(object, ...)
}
\arguments{
\item{object}{A `ps.cont` object}

\item{...}{Additional arguments.}
}
\description{
Displays a useful description of a `ps.cont` object.
}
